+++ ui/components/Signing/Signer/SignerLedger/SignerLedgerFrame.tsx
@@ -1,10 +1,10 @@
 import { SignOperationType } from "@tallyho/tally-background/redux-slices/signing"
 import React, { ReactElement, useCallback, useState } from "react"
-import { SigningFrameProps } from "../.."
 import { useBackgroundDispatch } from "../../../../hooks"
 import SharedButton from "../../../Shared/SharedButton"
 import SharedSlideUpMenu from "../../../Shared/SharedSlideUpMenu"
 import { useSigningLedgerState } from "../../../SignTransaction/useSigningLedgerState"
+import { SigningFrameProps } from "../../SigningFrame"
 import SignerLedgerConnect from "./SignerLedgerConnect"
 import SignerLedgerConnectionStatus from "./SignerLedgerConnectionStatus"
 import SignerLedgerSigning from "./SignerLedgerSigning"
