+++ ui/components/Signing/SigningData/SigningDataTransaction/SigningDataTransactionPanelCombined.tsx
@@ -1,11 +1,13 @@
+import { EnrichedEIP1559TransactionRequest } from "@tallyho/tally-background/services/enrichment"
 import React, { ReactElement } from "react"
-import { SigningDataTransactionProps } from "."
 import SigningDataTransactionDetailPanel from "./SigningDataTransactionDetailPanel"
 import SigningDataTransactionRawDataPanel from "./SigningDataTransactionRawDataPanel"
 
 export default function SigningDataTransactionPanelCombined({
   transactionRequest,
-}: SigningDataTransactionProps): ReactElement {
+}: {
+  transactionRequest: EnrichedEIP1559TransactionRequest
+}): ReactElement {
   return (
     <>
       <SigningDataTransactionDetailPanel
