+++ ui/components/Signing/SigningData/SigningDataTransaction/SigningDataTransactionSummary/SigningDataTransactionSummaryTransfer.tsx
@@ -6,8 +6,8 @@ import {
 import { enrichAssetAmountWithMainCurrencyValues } from "@tallyho/tally-background/redux-slices/utils/asset-utils"
 import { AssetTransfer } from "@tallyho/tally-background/services/enrichment"
 import React, { ReactElement } from "react"
-import {SigningDataTransactionSummaryBody} from "."
-import { SigningDataTransactionSummaryProps } from ".."
+import SigningDataTransactionSummaryBody from "./SigningDataTransactionSummaryBody"
+import { SigningDataTransactionSummaryProps } from "./SigningDataTransactionSummaryProps"
 import { useBackgroundSelector } from "../../../../../hooks"
 import SharedAddress from "../../../../Shared/SharedAddress"
