+++ ui/components/Signing/SigningData/SigningDataTransaction/SigningDataTransactionSummary/index.tsx
@@ -1,35 +1,10 @@
-import React, { ReactElement, ReactNode } from "react"
-import { SigningDataTransactionSummaryProps } from ".."
+import React, { ReactElement } from "react"
+import { SigningDataTransactionSummaryProps } from "./SigningDataTransactionSummaryProps"
 import SigningDataTransactionSummaryTransfer from "./SigningDataTransactionSummaryTransfer"
 import SigningDataTransactionSummarySpendApproval from "./SigningDataTransactionSummarySpendApproval"
 import SigningDataTransactionSummarySwapAsset from "./SigningDataTransactionSummarySwapAsset"
 import SigningDataTransactionSummaryDefault from "./SigningDataTransactionSummaryDefault"
 import SigningDataTransactionSummaryContractInteraction from "./SigningDataTransactionSummaryContractInteraction"
-
-export function SigningDataTransactionSummaryBody({
-  children,
-}: {
-  children: ReactNode
-}): ReactElement {
-  return (
-    <div>
-      {children}
-      <style jsx>
-        {`
-          div {
-            display: flex;
-            height: fit-content;
-            border-radius: 16px;
-            background-color: var(--hunter-green);
-            margin: 16px 0px;
-            flex-direction: column;
-            align-items: center;
-          }
-        `}
-      </style>
-    </div>
-  )
-}
 
 /**
  * Creates transaction type-specific summary blocks for use in parent
