+++ ui/pages/PersonalSign.tsx
@@ -10,6 +10,7 @@ import {
 } from "@tallyho/tally-background/redux-slices/signing"
 import { SignDataMessageType } from "@tallyho/tally-background/utils/signing"
 import { useHistory } from "react-router-dom"
+import { USE_UPDATED_SIGNING_UI } from "@tallyho/tally-background/features"
 import {
   useBackgroundDispatch,
   useBackgroundSelector,
@@ -17,8 +18,7 @@ import {
 } from "../hooks"
 import PersonalSignDetailPanel from "./PersonalSignDetailPanel"
 import SignTransactionContainer from "../components/SignTransaction/SignTransactionContainer"
-import {USE_UPDATED_SIGNING_UI} from "@tallyho/tally-background/features"
-import {Signing} from "../components/Signing"
+import Signing from "../components/Signing"
 
 const TITLE: Record<SignDataMessageType, string> = {
   [SignDataMessageType.EIP4361]: "Sign in with Ethereum",
